#199 refactor: icon naming 통일 showcase page
