+++ src/page/showcase/index/component/BannerSlider/BannerSlider.tsx
@@ -11,8 +11,8 @@ import 'slick-carousel/slick/slick.css';
 import { ReactNode, useMemo } from 'react';
 import Slider from 'react-slick';
 
-import LeftArrow from '@/common/asset/svg/arrow-left-white.svg?react';
-import RightArrow from '@/common/asset/svg/arrow-right-white.svg?react';
+import LeftArrow from '@/common/asset/svg/ic_arrow_left_white.svg?react';
+import RightArrow from '@/common/asset/svg/ic_arrow_right_white.svg?react';
 
 interface ArrowProps {
   onClick?: () => void;
