+++ src/page/showcase/index/component/ClubProfileCard/ClubProfileCard.tsx
@@ -5,7 +5,7 @@ import {
   imageStyle,
 } from '@/page/showcase/index/component/ClubProfileCard/ClubProfileCard.style';
 
-import defaultImage from '@/common/asset/svg/default_profile.svg';
+import defaultImage from '@/common/asset/svg/ic_default_profile.svg';
 import Heading from '@/common/component/Heading/Heading';
 import Text from '@/common/component/Text/Text';
